add: create navbar component to render in register page
